+++ src/Presentation/SmartStore.Web/Controllers/ProductController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
 using System.Web.Mvc;
+using System.Web.Routing;
 using SmartStore.Core.Domain.Catalog;
 using SmartStore.Core.Domain.Customers;
 using SmartStore.Core.Domain.Localization;
@@ -152,28 +153,34 @@ namespace SmartStore.Web.Controllers
 			if (!_storeMappingService.Authorize(product))
 				return HttpNotFound();
 
-			//visible individually?
+			// is product individually visible?
 			if (!product.VisibleIndividually)
 			{
-				//is this one an associated products?
+				// find parent grouped product
 				var parentGroupedProduct = _productService.GetProductById(product.ParentGroupedProductId);
-				if (parentGroupedProduct != null)
-				{
-					return RedirectToRoute("Product", new { SeName = parentGroupedProduct.GetSeName() });
-				}
-				else
-				{
+
+				if (parentGroupedProduct == null)
 					return HttpNotFound();
-				}
+
+				var routeValues = new RouteValueDictionary();
+				routeValues.Add("SeName", parentGroupedProduct.GetSeName());
+
+				// add query string parameters
+				Request.QueryString.AllKeys.Each(x => routeValues.Add(x, Request.QueryString[x]));
+
+				return RedirectToRoute("Product", routeValues);
 			}
 
 			var selectedAttributes = new NameValueCollection();
+			var attributesForProductId = 0;
+
+			if (product.ProductType == ProductType.GroupedProduct || (product.ProductType == ProductType.BundledProduct && product.BundlePerItemPricing))
+				attributesForProductId = 0;
+			else
+				attributesForProductId = product.Id;
 
 			// get selected attributes from query string
-			selectedAttributes.GetSelectedAttributes(
-				Request.QueryString,
-				_productAttributeParser.DeserializeQueryData(attributes),
-				product.ProductType == ProductType.BundledProduct && product.BundlePerItemPricing ? 0 : product.Id);
+			selectedAttributes.GetSelectedAttributes(Request.QueryString, _productAttributeParser.DeserializeQueryData(attributes),	attributesForProductId);
 
 			// prepare the view model
 			var model = _helper.PrepareProductDetailsPageModel(product, selectedAttributes: selectedAttributes, queryData: Request.QueryString);