+++ changelog.md
@@ -4,15 +4,20 @@
 
 ### Improvements
 * Added order message token for accepting third party email handover
+* ECB currency exchange rate provider now cross calculates rates based on euro rates
+* BeezUP: Exports the product weight
 
 ### New Features
-* Added option to skip shipping method selction in checkout process when only on shipping method is active
+* Added option to skip shipping method selection in checkout process when only one shipping method is active
 * Added options to capture salutation and title in addresses and customer info
+* Added projection to control the export of individually visible associated products
 
 ### Bugfixes
 * Currency wasn't displayed at shipping estimation
 * SKU, EAN, MPN of last attribute combination was exported for all combinations
 * GMC: Id should be unique when exporting attribute combinations as products
+* GMC: Attribute price adjustments were ignored when exporting attribute combinations as products
+* GMC: Associated products that are not individually visible are not exported anymore. GMC rejects them because the frontend redirects to the grouped product.
 
 
 ## SmartStore.NET 2.6